Fixed compiler warning on Visual Studio.
FossilOrigin-Name: 83495128c820e963c70d11c3196f81d8cf01f867

+++ tool/lemon.c
@@ -398,7 +398,7 @@ static int actioncmp(
     rc = ap1->x.rp->index - ap2->x.rp->index;
   }
   if( rc==0 ){
-    rc = ap2 - ap1;
+    rc = (int) (ap2 - ap1);
   }
   return rc;
 }
